+++ client/package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^12.8.3",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-router-dom": "^6.0.2",
     "react-scripts": "4.0.3",
     "socket.io-client": "^4.4.0",
     "web-vitals": "^1.1.1"
