+++ client/src/services/ErrorPage.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function ErrorPage() {
+    return (
+        <div>
+            <h2>ERROR! PAGE NOT FOUND!</h2>
+        </div>
+    )
+}
+
+export default ErrorPage
